refactor(modules): split the card pool out of modules/mod.rs
modules/mod.rs held 2411 lines of production code in one file: the
CardPool orchestrator (a single 1508-line inherent impl), the per-module
blocking AT worker, and the small types both sides share. mod.rs is now a
42-line facade; the rest is split by concern.

  protocol.rs       BridgeEvent / ModuleCmd / control channel aliases
  restart_policy.rs RestartMode + the two post-restart backoff constants
  slot.rs           SlotState, LifecycleState, PoolSlotView, slot queries
  worker.rs         the per-modem blocking AT loop
  pool/mod.rs       CardPool: bootstrap, retry, rescan, the select! loop
  pool/restart_cycle.rs  the [scheduled_restart] cycle
  pool/dispatch.rs  control commands, bridge events, outbound requests

Beyond relocation:

- run_module_loop's eight parameters (which needed an
  #[allow(clippy::too_many_arguments)]) become a ModuleWorker holding the
  loop state, with the URC handlers as methods. The pure helpers stay free
  functions so their existing tests keep working untouched.
- CardPool::run was 388 lines; the startup loop and the timer select! arm
  are now bootstrap_slots and on_timer_tick.
- The three identical 25-line worker-spawn blocks collapse into
  spawn_worker + worker_setup, and the two identical no-worker restart
  fallbacks into spawn_fallback_restart.
- Dropped two dead parameters: run_module_loop's _sms_enabled and
  handle_control_cmd's _resilience. Removing the former left CardPool's
  sms_handler field write-only, so it is now consumed by new() instead of
  stored; the store channel it cloned is kept alive by CardPool.store.

No behavior change. Test count is unchanged at 1268 passing, and the
externally imported surface (CardPool, ControlCmdSender,
ControlCmdReceiver) is re-exported from the facade, so no caller moved.
